Short tex translation

+++ include/global_data.h
@@ -623,8 +623,8 @@ int       firstspritelump, lastspritelump, numspritelumps;
 int       numtextures;
 texture_t **textures; // proff - 04/05/2000 removed static for OpenGL
 fixed_t   *textureheight; //needed for texture pegging (and TFE fix - killough)
-int       *flattranslation;             // for global animation
-int       *texturetranslation;
+short       *flattranslation;             // for global animation
+short       *texturetranslation;
 
 
 //******************************************************************************
